Add support to dump HP color space and transformation
The HP JPEG-LS implementation added 2 extensions to the ISO standard. One to store the color space information (ISO JPEG-LS stores this info in the SPIFF header) and info about the applied color transformation.
Normal JPEG-LS doesn't support color transformations, extended JPEG-LS does, but in standard JPEG-LS it is not stored in a APP8 markers segment.